refactor(custom-fields): move option label map into the app layer
